+++ src/RoadRegistry.BackOffice.Abstractions/RoadSegments/LinkRoadSegmentToStreetNameResponse.cs
@@ -1,3 +1,0 @@
-namespace RoadRegistry.BackOffice.Abstractions.RoadSegments;
-
-public sealed record LinkRoadSegmentToStreetNameResponse(Guid MessageId) : EndpointResponse;