+++ src/RoadRegistry.BackOffice.UI/init.sh
@@ -5,7 +5,7 @@ echo "" > /usr/share/nginx/html/env.js
 echo "window.API_VERSION = \"${API_VERSION}\"" >> /usr/share/nginx/html/env.js
 echo "window.API_ENDPOINT = \"${API_ENDPOINT}\"" >> /usr/share/nginx/html/env.js
 echo "window.API_OLDENDPOINT = \"${API_OLDENDPOINT}\"" >> /usr/share/nginx/html/env.js
-echo "window.featureToggles = {}" > /usr/share/nginx/html/env.js
+echo "window.featureToggles = {}" >> /usr/share/nginx/html/env.js
 echo "window.featureToggles.useFeatureCompare = \"${FeatureToggles__UseFeatureCompare}\"" >> /usr/share/nginx/html/env.js
 
 export DOLLAR=$