+++ src/RoadRegistry.BackOffice.ZipArchiveWriters/ExtractHost/IntegrationRoadNodesToZipArchiveWriter.cs
@@ -3,6 +3,7 @@ namespace RoadRegistry.BackOffice.ZipArchiveWriters.ExtractHost;
 using System.IO.Compression;
 using System.Text;
 using Be.Vlaanderen.Basisregisters.Shaperon;
+using Dbase.RoadNodes;
 using Editor.Schema;
 using Editor.Schema.RoadNodes;
 using Extensions;