+++ src/RoadRegistry.BackOffice.ZipArchiveWriters/ForEditor/RoadSegmentNumberedRoadAttributesToZipArchiveWriter.cs
@@ -3,8 +3,8 @@ namespace RoadRegistry.BackOffice.ZipArchiveWriters.ForEditor;
 using System.IO.Compression;
 using System.Text;
 using Be.Vlaanderen.Basisregisters.Shaperon;
+using Dbase.RoadSegments;
 using Editor.Schema;
-using Editor.Schema.RoadSegments;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.IO;