+++ src/RoadRegistry.BackOffice/FeatureToggles/UseOrganizationCreateFeatureToggle.cs
@@ -1,0 +1,3 @@
+namespace RoadRegistry.BackOffice.FeatureToggles;
+
+public sealed record UseOrganizationCreateFeatureToggle(bool FeatureEnabled) : ApplicationFeatureToggle(FeatureEnabled);