+++ src/RoadRegistry.Editor.Projections/GradeSeparatedJunctionRecordProjection.cs
@@ -6,6 +6,7 @@ using BackOffice;
 using BackOffice.Messages;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
 using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
+using Dbase.GradeSeparatedJuntions;
 using Microsoft.IO;
 using Schema;
 using Schema.GradeSeparatedJunctions;