+++ src/RoadRegistry.Wfs.ProjectionHost/Program.cs
@@ -164,6 +164,8 @@ public class Program
                         await WaitFor.SqlStreamStoreToBecomeAvailable(streamStore, logger).ConfigureAwait(false);
                         await migratorFactory.CreateMigrator(configuration, loggerFactory)
                             .MigrateAsync(CancellationToken.None).ConfigureAwait(false);
+
+                        Console.WriteLine("Started RoadRegistry.Wfs.ProjectionHost");
                         await eventProcessor.Resume(CancellationToken.None);
                     },
                     DistributedLockOptions.LoadFromConfiguration(configuration),